Specify Wifi channel for ESP-NOW transport
By default, the wifi channel on which ESP-Now is transmitting on ESP32 is channel 1. However in case a lot of wifi traffic is on this channel, it would be better to move to another wifi channel. A define ESP_NOW_WIFI_CHANNEL is added in config.h to specify the desired channel