I2S microphone distortion (#20)
The microphone (I2S) gives some strange distortion effect (noise) when it goes above maximum volume. This is caused by the 32-bit to 16-bit conversion of the audio, where the first 16 bits are simply ignored. This wrong, it should keep the maximimum 16 bit value 32767.

The distortion now has a more natural effect. But of course, the audio quality remains limited by the 8-bit transport protocol.

Probably it also improves the problems mentioned in
https://github.com/atomic14/esp32-walkie-talkie/issues/17
Remark that it should improve both ESP-now and TCP transmission.